// newProject/src/theme/colors/type.ts
export interface CommonColors {
  brand1: string;
  brand2: string;
  brand3: string;
  white: string;
  white1: string;
  white2: string;
  black: string;
  black1: string;
  black2: string;
  black40: string;
  blue: string;
  blue2: string;
  blue3: string;
  title: string;
  titleGray: string;
  textGray: string;
  gray1: string;
  gray2: string;
  red: string;
}

export enum Theme {
  LIGHT = 'light',
  DARK = 'dark'
}